docs: enrich converter comments for new contributors
