+++ Assets/Aramaa/OchibiChansConverterTool/Editor/Conversion/Avatar/OCTCostumeBlendShapeAdjuster.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 54660e09361b487c862f5368a286fe6f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
